chore: rename proximityhider package to proximity
